[arm] fix bootstrap to handle cli args correctly (details)
The initial bootstrap code for ARM was built to match gem5,
unfortuantely it seems like the way gem5 sets up it's stack is not
compatible with uclibc.

Not only was this breaking program execution for pydgin-arm, but gem5
will need to be modified as well to run apps compiled with uclibc.