[pydgin] remove @elidable from Debug.enabled
The elidable decorator was actually slightly hurting jit-nodebug because
at the header of every jit loop, the call to Debug.enabled was being made.
Because of the global_enabled flag, the optimizer can optimize this
function call away in jit-nodebug without the decorator.  Furthermore, it
was breaking jit-debug translation, which is working now.