[XPC-STATS] Adding task-size support for pcallrx
Now, one can use --task-size to alter the software strip-mining mechanism of pcallrx.
By doing so, the divergence generated and memory traces will be for a *task_size* chunk of
the body instead of the whole range executed as one loop.

This is achieved by modifying a0-a3. When pcallrx first invoked, pydgin makes note of regs
a0-a3. Then, pydgin set a1 to be the begnning of the task and a2 to be the end. Finally, as
JR's are being executed, pydgin keeps pushing a1 and a2 to the execution unit until there are
no more tasks left. With every JR, pydgin marks the begnning of a new "task/iteration."
Currently, there is no support for stride. Pydgin ignores it assuming it to be 1.

  Chunk-Range : [a1, a2)

  a0 : object pointer
  a1 : start index
  a2 : limit/max
  a3 : stride