Adding more robustness for task vs. data parallel sections
In this commit I detect a new parallel-section based on the first call to a
sync and bump the counter, clear state based on this logic. Most
app-kernels work with the exception of quickHull. The specific pattern is
the case where task calls a data-parallel section. NOT FIXING FOR NOW.