Split parallel regions for analysis!
Big commit which splits parallel regions into individual DAGs and analyzes
for similarity. The performance number is then obtained by the span of the
parallel region and the sum of serial isntructions. I also dump many other
useful stats and added sanity checks for the generated DAGs. Everything
works for tiny datasets and since, the DAGs are fully strict it should work
for larger datasets!