Fixes to minpc; mostly working
- Fixed the minpc based reconvergence scheme to not include the pc's when
not executing a parallel region -- this is important as otherwise, say a
non-master core is done with the parallel region, the core will enter back
the bthread spin-loop and the master core or other cores might be stuck
executing the spmd region but will never advance. NOTE: The implementation
is not elegant but works for now...

- changes to the linetrace which helps

- FIXME: I think there are application-level deadlocks and currently two
apps (mis/st) are not working yet. This commit also adds knn/hull back as
there was a typo in the doit scripts