Fixing a bug in stats collection!
Currently, I collect the unique and total number of instructions in the
recovergence manager which means I need to only collect stats for cores
that are active and are not stalling. If a core is waiting for other cores
or is at a barrier or did not get the fetch bandwidth allocated then it is
not going to execute. I had a bug where I collected stats regardless of the
states for each thread. This commit fixes this.