Initial commit for a plot workflow
The workflow is _super_ hacky but will still be useful and should suffice
for the moment. I haven't really set up the task dependencies correctly and
I should really be careful of what I am doing. However, this script will be
useful as I'll have to type very few commands